Create README.md
The purpose of the project is to showcase the student understanding on the concept of Linux scrip􀆟ng of wri􀆟ng a script that can be deployed into a network that can automa􀆟cally atack the systems in the network.
The script is set to be fully automated to perform all ac􀆟ons in a logical sequence except for op􀆟ons which require the user inputs. This script can be run by a SOC manager to test the SOC team’s vigilance.